+++ src/api/integrations/chatbot/dify/validate/dify.schema.ts
@@ -41,6 +41,8 @@ export const difySchema: JSONSchema7 = {
     keepOpen: { type: 'boolean' },
     debounceTime: { type: 'integer' },
     ignoreJids: { type: 'array', items: { type: 'string' } },
+    splitMessages: { type: 'boolean' },
+    timePerChar: { type: 'integer' },
   },
   required: ['enabled', 'botType', 'triggerType'],
   ...isNotEmpty('enabled', 'botType', 'triggerType'),
@@ -71,6 +73,8 @@ export const difySettingSchema: JSONSchema7 = {
     debounceTime: { type: 'integer' },
     ignoreJids: { type: 'array', items: { type: 'string' } },
     difyIdFallback: { type: 'string' },
+    splitMessages: { type: 'boolean' },
+    timePerChar: { type: 'integer' },
   },
   required: [
     'expire',
@@ -82,6 +86,8 @@ export const difySettingSchema: JSONSchema7 = {
     'keepOpen',
     'debounceTime',
     'ignoreJids',
+    'splitMessages',
+    'timePerChar',
   ],
   ...isNotEmpty(
     'expire',
@@ -93,6 +99,8 @@ export const difySettingSchema: JSONSchema7 = {
     'keepOpen',
     'debounceTime',
     'ignoreJids',
+    'splitMessages',
+    'timePerChar',
   ),
 };